refactor(package): test 코드 패키지 이동
